validate password
